create a cart folder where store components, install stripe
