+++ .env.example
@@ -12,3 +12,5 @@ NEXT_PUBLIC_CLOUDINARY_NAME=
 NEXT_PUBLIC_CLOUDINARY_API_KEY=
 NEXT_PUBLIC_CLOUDINARY_PRESET_SECRET=
 CLOUDINARY_API_SECRET=
+NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY=
+NEXT_PUBLIC_STRIPE_SECRET_KEY=
