+++ package.json
@@ -15,6 +15,8 @@
     "@mui/icons-material": "^5.15.13",
     "@mui/material": "^5.15.13",
     "@mui/material-nextjs": "^5.15.11",
+    "@stripe/react-stripe-js": "^2.7.0",
+    "@stripe/stripe-js": "^3.3.0",
     "bcryptjs": "^2.4.3",
     "cloudinary": "^2.0.3",
     "express-mongo-sanitize": "^2.2.0",
@@ -28,6 +30,7 @@
     "npm": "^10.5.0",
     "react": "^18",
     "react-dom": "^18",
+    "stripe": "^15.2.0",
     "zustand": "^4.5.2"
   },
   "devDependencies": {
